classing messages by consumers
