+++ helios-chain/x/hyperion/keeper/grpc_query.go
@@ -1,471 +1,7 @@
 package keeper
 
 import (
-	"context"
-	"fmt"
-	"strings"
-
-	"github.com/ethereum/go-ethereum/common"
-
-	"cosmossdk.io/errors"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-
-	"github.com/Helios-Chain-Labs/metrics"
-
 	"helios-core/helios-chain/x/hyperion/types"
 )
 
 var _ types.QueryServer = &Keeper{}
-
-const maxValsetRequestsReturned = 5
-const MaxResults = 100 // todo: impl pagination
-
-// [Used In Hyperion] Params queries the params of the hyperion module
-func (k *Keeper) Params(c context.Context, _ *types.QueryParamsRequest) (*types.QueryParamsResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	return &types.QueryParamsResponse{Params: *k.GetParams(sdk.UnwrapSDKContext(c))}, nil
-}
-
-// [Used In Hyperion] CurrentValset queries the CurrentValset of the hyperion module
-func (k *Keeper) CurrentValset(c context.Context, req *types.QueryCurrentValsetRequest) (*types.QueryCurrentValsetResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	return &types.QueryCurrentValsetResponse{Valset: k.GetCurrentValset(sdk.UnwrapSDKContext(c), req.HyperionId)}, nil
-}
-
-// [Used In Hyperion] ValsetRequest queries the ValsetRequest of the hyperion module
-func (k *Keeper) ValsetRequest(c context.Context, req *types.QueryValsetRequestRequest) (*types.QueryValsetRequestResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	return &types.QueryValsetRequestResponse{Valset: k.GetValset(sdk.UnwrapSDKContext(c), req.HyperionId, req.Nonce)}, nil
-}
-
-// [Not Used In Hyperion] ValsetConfirm queries the ValsetConfirm of the hyperion module
-func (k *Keeper) ValsetConfirm(c context.Context, req *types.QueryValsetConfirmRequest) (*types.QueryValsetConfirmResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	addr, err := sdk.AccAddressFromBech32(req.Address)
-	if err != nil {
-		return nil, errors.Wrap(sdkerrors.ErrInvalidRequest, "address invalid")
-	}
-
-	return &types.QueryValsetConfirmResponse{Confirm: k.GetValsetConfirm(sdk.UnwrapSDKContext(c), req.HyperionId, req.Nonce, addr)}, nil
-}
-
-// [Used In Hyperion] ValsetConfirmsByNonce queries the ValsetConfirmsByNonce of the hyperion module
-func (k *Keeper) ValsetConfirmsByNonce(c context.Context, req *types.QueryValsetConfirmsByNonceRequest) (*types.QueryValsetConfirmsByNonceResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	confirms := make([]*types.MsgValsetConfirm, 0)
-
-	k.IterateValsetConfirmByNonce(sdk.UnwrapSDKContext(c), req.HyperionId, req.Nonce, func(_ []byte, valset *types.MsgValsetConfirm) (stop bool) {
-		confirms = append(confirms, valset)
-
-		return false
-	})
-
-	return &types.QueryValsetConfirmsByNonceResponse{Confirms: confirms}, nil
-}
-
-// [Used In Hyperion] LastValsetRequests queries the LastValsetRequests of the hyperion module
-func (k *Keeper) LastValsetRequests(c context.Context, req *types.QueryLastValsetRequestsRequest) (*types.QueryLastValsetRequestsResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	valReq := k.GetValsets(sdk.UnwrapSDKContext(c), req.HyperionId)
-	valReqLen := len(valReq)
-	retLen := 0
-
-	if valReqLen < maxValsetRequestsReturned {
-		retLen = valReqLen
-	} else {
-		retLen = maxValsetRequestsReturned
-	}
-
-	return &types.QueryLastValsetRequestsResponse{Valsets: valReq[0:retLen]}, nil
-}
-
-// [Used In Hyperion] LastPendingValsetRequestByAddr queries the LastPendingValsetRequestByAddr of the hyperion module
-func (k *Keeper) LastPendingValsetRequestByAddr(c context.Context, req *types.QueryLastPendingValsetRequestByAddrRequest) (*types.QueryLastPendingValsetRequestByAddrResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	addr, err := sdk.AccAddressFromBech32(req.Address)
-	if err != nil {
-		metrics.ReportFuncError(k.svcTags)
-		return nil, errors.Wrap(sdkerrors.ErrInvalidRequest, "address invalid")
-	}
-
-	pendingValsetReq := make([]*types.Valset, 0)
-	k.IterateValsets(sdk.UnwrapSDKContext(c), func(_ []byte, val *types.Valset) bool {
-		if val.HyperionId != req.HyperionId {
-			// return false to continue the loop
-			return false
-		}
-		// foundConfirm is true if the operatorAddr has signed the valset we are currently looking at
-		foundConfirm := k.GetValsetConfirm(sdk.UnwrapSDKContext(c), val.HyperionId, val.Nonce, addr) != nil
-		// if this valset has NOT been signed by operatorAddr, store it in pendingValsetReq
-		// and exit the loop
-		if !foundConfirm {
-			pendingValsetReq = append(pendingValsetReq, val)
-		}
-		// if we have more than 100 unconfirmed requests in
-		// our array we should exit, TODO pagination
-		if len(pendingValsetReq) > 100 {
-			return true
-		}
-		// return false to continue the loop
-		return false
-	})
-
-	return &types.QueryLastPendingValsetRequestByAddrResponse{Valsets: pendingValsetReq}, nil
-}
-
-// [Used In Hyperion] BatchFees queries the batch fees from unbatched pool
-func (k *Keeper) BatchFees(c context.Context, req *types.QueryBatchFeeRequest) (*types.QueryBatchFeeResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	return &types.QueryBatchFeeResponse{BatchFees: k.GetAllBatchFees(sdk.UnwrapSDKContext(c), req.HyperionId)}, nil
-}
-
-// [Used In Hyperion] LastPendingBatchRequestByAddr queries the LastPendingBatchRequestByAddr of the hyperion module
-func (k *Keeper) LastPendingBatchRequestByAddr(c context.Context, req *types.QueryLastPendingBatchRequestByAddrRequest) (*types.QueryLastPendingBatchRequestByAddrResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	addr, err := sdk.AccAddressFromBech32(req.Address)
-	if err != nil {
-		metrics.ReportFuncError(k.svcTags)
-		return nil, errors.Wrap(sdkerrors.ErrInvalidRequest, "address invalid")
-	}
-
-	var pendingBatchReq *types.OutgoingTxBatch
-	k.IterateOutgoingTXBatches(sdk.UnwrapSDKContext(c), req.HyperionId, func(_ []byte, batch *types.OutgoingTxBatch) (stop bool) {
-		foundConfirm := k.GetBatchConfirm(sdk.UnwrapSDKContext(c), batch.HyperionId, batch.BatchNonce, common.HexToAddress(batch.TokenContract), addr) != nil
-		if !foundConfirm {
-			pendingBatchReq = batch
-			return true
-		}
-
-		return false
-	})
-
-	return &types.QueryLastPendingBatchRequestByAddrResponse{Batch: pendingBatchReq}, nil
-}
-
-// [Used In Hyperion] OutgoingTxBatches queries the OutgoingTxBatches of the hyperion module
-func (k *Keeper) OutgoingTxBatches(c context.Context, req *types.QueryOutgoingTxBatchesRequest) (*types.QueryOutgoingTxBatchesResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	batches := make([]*types.OutgoingTxBatch, 0)
-	k.IterateOutgoingTXBatches(sdk.UnwrapSDKContext(c), req.HyperionId, func(_ []byte, batch *types.OutgoingTxBatch) bool {
-		batches = append(batches, batch)
-		return len(batches) == MaxResults
-	})
-
-	return &types.QueryOutgoingTxBatchesResponse{Batches: batches}, nil
-}
-
-// [Not Used In Hyperion] BatchRequestByNonce queries the BatchRequestByNonce of the hyperion module
-func (k *Keeper) BatchRequestByNonce(c context.Context, req *types.QueryBatchRequestByNonceRequest) (*types.QueryBatchRequestByNonceResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	if err := types.ValidateEthAddress(req.ContractAddress); err != nil {
-		metrics.ReportFuncError(k.svcTags)
-		return nil, errors.Wrap(sdkerrors.ErrUnknownRequest, err.Error())
-	}
-
-	foundBatch := k.GetOutgoingTXBatch(sdk.UnwrapSDKContext(c), common.HexToAddress(req.ContractAddress), req.Nonce, req.HyperionId)
-	if foundBatch == nil {
-		metrics.ReportFuncError(k.svcTags)
-		return nil, errors.Wrap(sdkerrors.ErrUnknownRequest, "Can not find tx batch")
-	}
-
-	return &types.QueryBatchRequestByNonceResponse{Batch: foundBatch}, nil
-}
-
-// [Used In Hyperion] BatchConfirms returns the batch confirmations by nonce and token contract
-func (k *Keeper) BatchConfirms(c context.Context, req *types.QueryBatchConfirmsRequest) (*types.QueryBatchConfirmsResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	confirms := make([]*types.MsgConfirmBatch, 0)
-	k.IterateBatchConfirmByNonceAndTokenContract(sdk.UnwrapSDKContext(c), req.HyperionId, req.Nonce, common.HexToAddress(req.ContractAddress),
-		func(_ []byte, batch *types.MsgConfirmBatch) (stop bool) {
-			confirms = append(confirms, batch)
-			return false
-		})
-
-	return &types.QueryBatchConfirmsResponse{Confirms: confirms}, nil
-}
-
-// [Used In Hyperion] LastEventByAddr returns the last event for the given validator address, this allows eth oracles to figure out where they left off
-func (k *Keeper) LastEventByAddr(c context.Context, req *types.QueryLastEventByAddrRequest) (*types.QueryLastEventByAddrResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	ctx := sdk.UnwrapSDKContext(c)
-	var ret types.QueryLastEventByAddrResponse
-
-	k.Logger(ctx).Info("LastEventByAddr")
-
-	addr, err := sdk.AccAddressFromBech32(req.Address)
-	if err != nil {
-		metrics.ReportFuncError(k.svcTags)
-		return nil, errors.Wrap(sdkerrors.ErrInvalidAddress, req.Address)
-	}
-
-	validator, found := k.GetOrchestratorValidator(ctx, req.HyperionId, addr)
-	if !found {
-		metrics.ReportFuncError(k.svcTags)
-		return nil, errors.Wrap(types.ErrUnknown, "address")
-	}
-
-	lastClaimEvent := k.GetLastEventByValidatorAndHyperionId(ctx, req.HyperionId, validator)
-	if lastClaimEvent.EthereumEventNonce == 0 && lastClaimEvent.EthereumEventHeight == 0 {
-		// if hyperion happens to query too early without a bonded validator even existing setup the base event
-		lowestObservedNonce := k.GetLastObservedEventNonce(ctx, req.HyperionId)
-		blockHeight := k.GetLastObservedEthereumBlockHeight(ctx, req.HyperionId).EthereumBlockHeight
-
-		k.setLastEventByValidatorAndHyperionId(
-			ctx,
-			req.HyperionId,
-			validator,
-			lowestObservedNonce,
-			blockHeight,
-		)
-		lastClaimEvent = k.GetLastEventByValidatorAndHyperionId(ctx, req.HyperionId, validator)
-	}
-
-	ret.LastClaimEvent = &lastClaimEvent
-
-	return &ret, nil
-}
-
-// [Not Used In Hyperion] DenomToERC20 queries the Cosmos Denom that maps to an Ethereum ERC20
-func (k *Keeper) DenomToERC20(c context.Context, req *types.QueryDenomToERC20Request) (*types.QueryDenomToERC20Response, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	ctx := sdk.UnwrapSDKContext(c)
-	cosmosOriginated, erc20, err := k.DenomToERC20Lookup(ctx, req.Denom, req.HyperionId)
-
-	var ret types.QueryDenomToERC20Response
-	ret.Erc20 = erc20.Hex()
-	ret.CosmosOriginated = cosmosOriginated
-
-	return &ret, err
-}
-
-// [Not Used In Hyperion] ERC20ToDenom queries the ERC20 contract that maps to an Ethereum ERC20 if any
-func (k *Keeper) ERC20ToDenom(c context.Context, req *types.QueryERC20ToDenomRequest) (*types.QueryERC20ToDenomResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	ctx := sdk.UnwrapSDKContext(c)
-	cosmosOriginated, name := k.ERC20ToDenomLookup(ctx, common.HexToAddress(req.Erc20), req.HyperionId)
-
-	var ret types.QueryERC20ToDenomResponse
-	ret.Denom = name
-	ret.CosmosOriginated = cosmosOriginated
-
-	return &ret, nil
-}
-
-// [Not Used In Hyperion]
-func (k *Keeper) GetDelegateKeyByValidator(c context.Context, req *types.QueryDelegateKeysByValidatorAddress) (*types.QueryDelegateKeysByValidatorAddressResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	ctx := sdk.UnwrapSDKContext(c)
-
-	valAddress, err := sdk.ValAddressFromBech32(req.ValidatorAddress)
-	if err != nil {
-		metrics.ReportFuncError(k.svcTags)
-		return nil, err
-	}
-
-	valAccountAddr := sdk.AccAddress(valAddress.Bytes())
-	keys := k.GetOrchestratorAddresses(ctx, req.HyperionId)
-
-	for _, key := range keys {
-		senderAddr, err := sdk.AccAddressFromBech32(key.Sender)
-		if err != nil {
-			metrics.ReportFuncError(k.svcTags)
-			return nil, err
-		}
-		if valAccountAddr.Equals(senderAddr) {
-			return &types.QueryDelegateKeysByValidatorAddressResponse{EthAddress: key.EthAddress, OrchestratorAddress: key.Orchestrator}, nil
-		}
-	}
-
-	metrics.ReportFuncError(k.svcTags)
-	return nil, errors.Wrap(types.ErrInvalid, "No validator")
-}
-
-// [Not Used In Hyperion]
-func (k *Keeper) GetDelegateKeyByOrchestrator(c context.Context, req *types.QueryDelegateKeysByOrchestratorAddress) (*types.QueryDelegateKeysByOrchestratorAddressResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	ctx := sdk.UnwrapSDKContext(c)
-	keys := k.GetOrchestratorAddresses(ctx, req.HyperionId)
-
-	_, err := sdk.AccAddressFromBech32(req.OrchestratorAddress)
-	if err != nil {
-		metrics.ReportFuncError(k.svcTags)
-		return nil, err
-	}
-
-	for _, key := range keys {
-		if req.OrchestratorAddress == key.Orchestrator {
-			return &types.QueryDelegateKeysByOrchestratorAddressResponse{ValidatorAddress: key.Sender, EthAddress: key.EthAddress}, nil
-		}
-	}
-
-	metrics.ReportFuncError(k.svcTags)
-	return nil, errors.Wrap(types.ErrInvalid, "No validator")
-}
-
-// [Used In Hyperion]
-func (k *Keeper) GetDelegateKeyByEth(c context.Context, req *types.QueryDelegateKeysByEthAddress) (*types.QueryDelegateKeysByEthAddressResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	ctx := sdk.UnwrapSDKContext(c)
-
-	keys := k.GetOrchestratorAddresses(ctx, req.HyperionId)
-	fmt.Println("keys: ", keys)
-	if err := types.ValidateEthAddress(req.EthAddress); err != nil {
-		return nil, errors.Wrap(err, "invalid eth address")
-	}
-
-	for _, key := range keys {
-		if req.EthAddress == key.EthAddress {
-			return &types.QueryDelegateKeysByEthAddressResponse{
-				ValidatorAddress:    key.Sender,
-				OrchestratorAddress: key.Orchestrator}, nil
-		}
-	}
-
-	metrics.ReportFuncError(k.svcTags)
-	return nil, errors.Wrap(types.ErrInvalid, "No validator")
-}
-
-// [Not Used In Hyperion]
-func (k *Keeper) GetPendingSendToChain(c context.Context, req *types.QueryPendingSendToChain) (*types.QueryPendingSendToChainResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	ctx := sdk.UnwrapSDKContext(c)
-	batches := k.GetOutgoingTxBatches(ctx, req.HyperionId)
-	unbatchedTx := k.GetPoolTransactions(ctx, req.HyperionId)
-	senderAddress := req.SenderAddress
-
-	res := &types.QueryPendingSendToChainResponse{}
-	res.TransfersInBatches = make([]*types.OutgoingTransferTx, 0)
-	res.UnbatchedTransfers = make([]*types.OutgoingTransferTx, 0)
-
-	for _, batch := range batches {
-		for _, tx := range batch.Transactions {
-			if tx.Sender == senderAddress {
-				res.TransfersInBatches = append(res.TransfersInBatches, tx)
-			}
-		}
-	}
-
-	for _, tx := range unbatchedTx {
-		if strings.EqualFold(tx.Sender, senderAddress) {
-			res.UnbatchedTransfers = append(res.UnbatchedTransfers, tx)
-
-		}
-	}
-
-	return res, nil
-}
-
-// [Not Used In Hyperion]
-func (k *Keeper) GetAllPendingSendToChain(c context.Context, req *types.QueryAllPendingSendToChainRequest) (*types.QueryAllPendingSendToChainResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	ctx := sdk.UnwrapSDKContext(c)
-	batches := k.GetAllOutgoingTxBatches(ctx)
-	unbatchedTx := k.GetAllPoolTransactions(ctx)
-
-	res := &types.QueryAllPendingSendToChainResponse{}
-	res.TransfersInBatches = make([]*types.OutgoingTransferTx, 0)
-	res.UnbatchedTransfers = make([]*types.OutgoingTransferTx, 0)
-
-	for _, batch := range batches {
-		res.TransfersInBatches = append(res.TransfersInBatches, batch.Transactions...)
-	}
-	res.UnbatchedTransfers = append(res.UnbatchedTransfers, unbatchedTx...)
-
-	return res, nil
-}
-
-// [Not Used In Hyperion]
-func (k *Keeper) HyperionModuleState(c context.Context, req *types.QueryModuleStateRequest) (*types.QueryModuleStateResponse, error) {
-	c, doneFn := metrics.ReportFuncCallAndTimingCtx(c, k.grpcTags)
-	defer doneFn()
-
-	ctx := sdk.UnwrapSDKContext(c)
-	state := ExportGenesis(ctx, *k)
-
-	res := &types.QueryModuleStateResponse{
-		State: &state,
-	}
-
-	return res, nil
-}
-
-// [Not Used In Hyperion]
-func (k *Keeper) MissingHyperionNonces(
-	c context.Context,
-	req *types.MissingNoncesRequest,
-) (*types.MissingNoncesResponse, error) {
-	ctx := sdk.UnwrapSDKContext(c)
-	var res []string
-
-	bondedValidators, err := k.StakingKeeper.GetBondedValidatorsByPower(ctx)
-	if err != nil {
-		return nil, err
-	}
-	for i := range bondedValidators {
-		val, _ := sdk.ValAddressFromBech32(bondedValidators[i].GetOperator())
-
-		ev := k.GetLastEventByValidatorAndHyperionId(ctx, req.HyperionId, val)
-		if ev.EthereumEventNonce == 0 && ev.EthereumEventHeight == 0 {
-			res = append(res, bondedValidators[i].GetOperator())
-		}
-	}
-
-	return &types.MissingNoncesResponse{OperatorAddresses: res}, nil
-}
-
-// [Not Used In Hyperion]
-func (k *Keeper) GetHyperionIdFromChainId(c context.Context, req *types.QueryGetHyperionIdFromChainIdRequest) (*types.QueryGetHyperionIdFromChainIdResponse, error) {
-	ctx := sdk.UnwrapSDKContext(c)
-	params := k.GetParams(ctx)
-
-	for _, counterpartyChainParam := range params.CounterpartyChainParams {
-		if counterpartyChainParam.BridgeChainId == req.ChainId {
-			return &types.QueryGetHyperionIdFromChainIdResponse{
-				HyperionId: counterpartyChainParam.HyperionId,
-			}, nil
-		}
-	}
-
-	return nil, errors.Wrap(types.ErrDuplicate, "BridgeChainId not found")
-}
