--- helios-chain/x/hyperion/keeper/grpc_query_attestation.go
@@ -1,28 +1,0 @@
-package keeper
-
-import (
-	"context"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-
-	"helios-core/helios-chain/x/hyperion/types"
-)
-
-func (k Keeper) Attestation(c context.Context, req *types.QueryAttestationRequest) (*types.QueryAttestationResponse, error) {
-	if req == nil {
-		return nil, status.Error(codes.InvalidArgument, "empty request")
-	}
-
-	ctx := sdk.UnwrapSDKContext(c)
-
-	att := k.GetAttestation(ctx, req.HyperionId, req.Nonce, req.ClaimHash)
-	if att == nil {
-		return nil, status.Error(codes.NotFound, "attestation not found")
-	}
-
-	return &types.QueryAttestationResponse{
-		Attestation: att,
-	}, nil
-}
